Add basic crash reporting to near-cli
This requires adding a workaround to spawn a child process, since
mixpanel does not handle this case properly.

There is an existing limitation that if a user did not opt in to
tracking, the crash report will not be sent. We may want to ask an
additional qustion for this case, since there are scenarios where people
are willing to send an individual crash report but not opt in to general
tracking.

There is a known limitation that this implementation does not log the
environment that was used.